还是把int()加回来了。。

+++ graph_segmentation.py
@@ -44,9 +44,9 @@ class GraphSegmentationEuclideanRGB(GraphSegmentationDistance):
 
     # Evaluate the distance given 2 nodes.
     def __call__(self, n, m):
-        dr = n.r - m.r
-        dg = n.g - m.g
-        db = n.b - m.b
+        dr = int(n.r) - int(m.r)
+        dg = int(n.g) - int(m.g)
+        db = int(n.b) - int(m.b)
         return math.sqrt(dr * dr + dg * dg + db * db) / self.__D
 
 
@@ -60,12 +60,12 @@ class GraphSegmentationEuclideanRGBFlowAngle(GraphSegmentationDistance):
 
     def __call__(self, n, m):
         assert (len(self._weights) == 2)
-        #dr = int(n.r) - int(m.r)
-        #dg = int(n.g) - int(m.g)
-        #db = int(n.b) - int(m.b)
-        dr = (n.r) - (m.r)
-        dg = (n.g) - (m.g)
-        db = (n.b) - (m.b)
+        dr = int(n.r) - int(m.r)
+        dg = int(n.g) - int(m.g)
+        db = int(n.b) - int(m.b)
+        #dr = (n.r) - (m.r)
+        #dg = (n.g) - (m.g)
+        #db = (n.b) - (m.b)
         if n.t == m.t:
             # Use flow within frames.
             n_f = math.sqrt(n.fx * n.fx + n.fy * n.fy)
@@ -77,10 +77,10 @@ class GraphSegmentationEuclideanRGBFlowAngle(GraphSegmentationDistance):
             a = math.acos(cos_a)
             pi = math.pi + 1e-4
             assert (0 <= a <= pi)
-            return self._weights[0] * math.sqrt(dr * dr + dg * dg + db * db) / float(self.__D) + self._weights[
+            return self._weights[0] * math.sqrt(int(dr * dr) + int(dg * dg) + int(db * db)) / float(self.__D) + self._weights[
                                                                                                      1] * a / pi
         else:
-            return math.sqrt(dr * dr + dg * dg + db * db) / self.__D
+            return math.sqrt(int(dr * dr) + int(dg * dg) + int(db * db)) / self.__D
 
 
 class GraphSegmentationMagic:
